Major rework of HTML page. A couple of minor bug fixes.
I'm including a PNG (PiServed_Web_Interface_as_of_20160318.png) which is
an annotated screenshot of the actual web page. I figure this way, even
for people that do not have a Pi, they can at least see what it is
they're missing. (In some small way.) Watch for a video demo soon! :)